refactor(share): collapse the ShareUtils.js/ + utils/share double tree
utils/ShareUtils.js/ (old) and utils/share/ (new) both implemented
stats share text, canvas watermarks and collage layouts, and the mixed
index.js exported an unpredictable blend: explicit re-exports (new
implementations) shadowed star exports (old), except
generateStatsShareText where the old version silently won and the two
copies had drifted apart in format.

utils/share/ is now the only tree. StatsTextGenerator.js takes over
the old generateStatsShareText body verbatim - the version users
currently see - locked by a new format test, with one fix: lines are
joined with a real newline instead of the literal backslash-n both
copies emitted into the clipboard text. The old image-processing and
collage-layout modules were unreachable (shadowed or unimported) and
are deleted with the rest of ShareUtils.js/.
