--- src/utils/ShareUtils.js/index.js
@@ -1,15 +1,0 @@
-/**
- * ShareUtils - Utilities for sharing stats and images
- */
-
-// Re-export all functions from the sub-modules
-export * from './stats-share.js';
-export * from './image-processing.js';
-export * from './collage-layouts.js';
-
-// Re-export from share/ directory
-export { copyTextToClipboard, copyImageToClipboard, saveImageAsFile } from '../share/ClipboardUtils.js';
-export { getCollageLayout, generateCollage } from '../share/CollageGenerator.js';
-export { shareToSocial } from '../share/SocialMediaShare.js';
-export { generateShareablePhoto } from '../share/ImageProcessingUtils.js';
-export { generateStatsImage } from '../share/StatsImageGenerator.js';
